Reorganise the libcamera-apps source tree.
The source tree is becoming a bit crowded to use a flat structure now. This
commit reorganises the tree into the following directories:

apps - Top level applications
core - Core libcamera_app files
encoder - Various stream encoders
image - Still image helpers
output - Output network/file sink helpers
preview - EGL/DRM and null preview helpers
