+++ core/frame_info.hpp
