+++ core/libcamera_app.hpp
