+++ core/video_options.hpp
