+++ encoder/encoder.cpp
