+++ encoder/encoder.hpp
@@ -9,7 +9,7 @@
 
 #include <functional>
 
-#include "video_options.hpp"
+#include "core/video_options.hpp"
 
 typedef std::function<void(void *)> InputDoneCallback;
 typedef std::function<void(void *, size_t, int64_t, bool)> OutputReadyCallback;
