+++ encoder/mjpeg_encoder.cpp
