+++ encoder/mjpeg_encoder.hpp
