+++ encoder/null_encoder.cpp
