+++ image/bmp.cpp
@@ -10,7 +10,7 @@
 
 #include <libcamera/formats.h>
 
-#include "still_options.hpp"
+#include "core/still_options.hpp"
 
 struct ImageHeader
 {
