+++ output/circular_output.cpp
