+++ output/circular_output.hpp
