+++ output/net_output.cpp
