+++ output/net_output.hpp
